feat(Table): extract link to log line code
